# swagger.yml
swagger: "2.0"

info:
  description: "E-commerce REST API with full CRUD operations"
  version: "1.0.0"
  title: "E-commerce REST API"
  contact:
    name: "API Support"

schemes: 
  - http

host: localhost:4000

basePath: "/"

consumes:
  - application/json
produces:
  - application/json

paths:
  /auth/register:
    post:
      summary: "Register a new user"
      description: "Create a new user account with email and password"
      parameters:
        - in: body
          name: body
          description: "User registration details"
          required: true
          schema:
            type: object
            required:
              - email
              - password
              - firstName
              - lastName
            properties:
              email:
                type: string
                example: "user@example.com"
              password:
                type: string
                example: "password123"
              firstName:
                type: string
                example: "John"
              lastName:
                type: string
                example: "Doe"
      responses:
        201:
          description: "User successfully registered"
          schema:
            type: object
            properties:
              id:
                type: integer
              email:
                type: string
              firstName:
                type: string
              lastName:
                type: string
        400:
          description: "Bad request - missing required fields"
        409:
          description: "Email already in use"
        500:
          description: "Internal server error"

  /auth/login:
    post:
      summary: "Login to user account"
      description: "Authenticate user with email and password"
      parameters:
        - in: body
          name: body
          description: "User login credentials"
          required: true
          schema:
            type: object
            required:
              - username
              - password
            properties:
              username:
                type: string
                description: "User email"
                example: "user@example.com"
              password:
                type: string
                example: "password123"
      responses:
        200:
          description: "Login successful"
          schema:
            type: object
            properties:
              id:
                type: integer
              email:
                type: string
              firstName:
                type: string
              lastName:
                type: string
        401:
          description: "Incorrect username or password"
        500:
          description: "Internal server error"

  /auth/logout:
    post:
      summary: "Logout from user account"
      description: "End the current session"
      responses:
        200:
          description: "Successfully logged out"
          schema:
            type: object
            properties:
              message:
                type: string
        500:
          description: "Internal server error"

  /users/{userId}:
    get:
      summary: "Get user record"
      description: "Retrieve user information by ID"
      parameters:
        - in: path
          name: userId
          type: integer
          required: true
          description: "User ID"
      responses:
        200:
          description: "User found"
          schema:
            type: object
            properties:
              id:
                type: integer
              email:
                type: string
              firstName:
                type: string
              lastName:
                type: string
        404:
          description: "User not found"
        500:
          description: "Internal server error"

    put:
      summary: "Update user record"
      description: "Update user information"
      parameters:
        - in: path
          name: userId
          type: integer
          required: true
          description: "User ID"
        - in: body
          name: body
          description: "User data to update"
          required: true
          schema:
            type: object
            properties:
              firstName:
                type: string
              lastName:
                type: string
              password:
                type: string
      responses:
        200:
          description: "User updated successfully"
          schema:
            type: object
        500:
          description: "Internal server error"

    delete:
      summary: "Delete user record"
      description: "Delete a user account"
      parameters:
        - in: path
          name: userId
          type: integer
          required: true
          description: "User ID"
      responses:
        200:
          description: "User deleted successfully"
          schema:
            type: object
        404:
          description: "User not found"
        500:
          description: "Internal server error"

  /products:
    get:
      summary: "List all products"
      description: "Retrieve all available products"
      parameters:
        - in: query
          name: limit
          type: integer
          description: "Limit number of results"
        - in: query
          name: offset
          type: integer
          description: "Offset for pagination"
      responses:
        200:
          description: "Product list retrieved successfully"
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
                price:
                  type: integer
                description:
                  type: string
        500:
          description: "Internal server error"

    post:
      summary: "Create a new product"
      description: "Add a new product to the catalog"
      parameters:
        - in: body
          name: body
          description: "Product details"
          required: true
          schema:
            type: object
            required:
              - name
              - price
              - description
            properties:
              name:
                type: string
                example: "Product Name"
              price:
                type: integer
                example: 1999
              description:
                type: string
                example: "Product description"
      responses:
        201:
          description: "Product created successfully"
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              price:
                type: integer
              description:
                type: string
        400:
          description: "Bad request - missing required fields"
        500:
          description: "Internal server error"

  /products/{productId}:
    get:
      summary: "Get product by ID"
      description: "Retrieve a specific product"
      parameters:
        - in: path
          name: productId
          type: integer
          required: true
          description: "Product ID"
      responses:
        200:
          description: "Product found"
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              price:
                type: integer
              description:
                type: string
        404:
          description: "Product not found"
        500:
          description: "Internal server error"

    put:
      summary: "Update product"
      description: "Update product information"
      parameters:
        - in: path
          name: productId
          type: integer
          required: true
          description: "Product ID"
        - in: body
          name: body
          description: "Product data to update"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              price:
                type: integer
              description:
                type: string
      responses:
        200:
          description: "Product updated successfully"
          schema:
            type: object
        404:
          description: "Product not found"
        500:
          description: "Internal server error"

    delete:
      summary: "Delete product"
      description: "Remove a product from the catalog"
      parameters:
        - in: path
          name: productId
          type: integer
          required: true
          description: "Product ID"
      responses:
        200:
          description: "Product deleted successfully"
          schema:
            type: object
        404:
          description: "Product not found"
        500:
          description: "Internal server error"

  /carts/mine:
    get:
      summary: "Get user's cart"
      description: "Retrieve the current user's shopping cart"
      responses:
        200:
          description: "Cart retrieved successfully"
          schema:
            type: object
            properties:
              id:
                type: integer
              userId:
                type: integer
              items:
                type: array
                items:
                  type: object
        404:
          description: "Cart not found"
        500:
          description: "Internal server error"

    post:
      summary: "Create a new cart"
      description: "Create a new shopping cart for the user"
      responses:
        200:
          description: "Cart created successfully"
          schema:
            type: object
            properties:
              id:
                type: integer
              userId:
                type: integer
        500:
          description: "Internal server error"

    put:
      summary: "Get cart details"
      description: "Retrieve cart details by cart ID"
      responses:
        200:
          description: "Cart retrieved successfully"
          schema:
            type: object
        500:
          description: "Internal server error"

  /carts/mine/items:
    post:
      summary: "Add item to cart"
      description: "Add a product to the user's cart"
      parameters:
        - in: body
          name: body
          description: "Cart item details"
          required: true
          schema:
            type: object
            required:
              - productId
              - qty
            properties:
              productId:
                type: integer
              qty:
                type: integer
      responses:
        200:
          description: "Item added to cart successfully"
          schema:
            type: object
        400:
          description: "Bad request - missing required fields"
        500:
          description: "Internal server error"

  /carts/mine/items/{cartItemId}:
    put:
      summary: "Update cart item"
      description: "Update quantity of an item in the cart"
      parameters:
        - in: path
          name: cartItemId
          type: integer
          required: true
          description: "Cart item ID"
        - in: body
          name: body
          description: "Updated cart item data"
          required: true
          schema:
            type: object
            properties:
              qty:
                type: integer
      responses:
        200:
          description: "Cart item updated successfully"
          schema:
            type: object
        500:
          description: "Internal server error"

    delete:
      summary: "Remove item from cart"
      description: "Delete an item from the user's cart"
      parameters:
        - in: path
          name: cartItemId
          type: integer
          required: true
          description: "Cart item ID"
      responses:
        200:
          description: "Item removed from cart successfully"
          schema:
            type: object
        404:
          description: "Cart item not found"
        500:
          description: "Internal server error"

  /carts/mine/checkout:
    post:
      summary: "Checkout cart"
      description: "Process cart checkout and create an order"
      parameters:
        - in: body
          name: body
          description: "Checkout information"
          required: true
          schema:
            type: object
            properties:
              cartId:
                type: integer
              paymentInfo:
                type: object
                properties:
                  id:
                    type: string
      responses:
        200:
          description: "Order created successfully"
          schema:
            type: object
            properties:
              id:
                type: integer
              total:
                type: integer
              status:
                type: string
              userId:
                type: integer
        400:
          description: "Bad request or payment failed"
        500:
          description: "Internal server error"

  /orders:
    get:
      summary: "List user's orders"
      description: "Retrieve all orders for the authenticated user"
      responses:
        200:
          description: "Orders retrieved successfully"
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                total:
                  type: integer
                status:
                  type: string
                userId:
                  type: integer
                created:
                  type: string
        500:
          description: "Internal server error"

  /orders/{orderId}:
    get:
      summary: "Get order by ID"
      description: "Retrieve a specific order"
      parameters:
        - in: path
          name: orderId
          type: integer
          required: true
          description: "Order ID"
      responses:
        200:
          description: "Order found"
          schema:
            type: object
            properties:
              id:
                type: integer
              total:
                type: integer
              status:
                type: string
              userId:
                type: integer
              items:
                type: array
                items:
                  type: object
        404:
          description: "Order not found"
        500:
          description: "Internal server error"